create KpmManager and migrate PluginData class
